refactor few cyclic dependencies
