+++ crates/proof-of-sql/src/base/arrow/mod.rs
@@ -1,0 +1,9 @@
+pub mod arrow_array_to_column_conversion;
+pub mod record_batch_utility;
+pub mod record_batch_conversion;
+pub mod record_batch_errors;
+pub mod owned_and_arrow_conversions;
+pub mod scalar_and_i256_conversions;
+
+#[cfg(test)]
+pub mod owned_and_arrow_conversions_test;
