+++ crates/proof-of-sql/src/base/commitment/table_commitment.rs
@@ -2,18 +2,13 @@ use super::{
     committable_column::CommittableColumn, AppendColumnCommitmentsError, ColumnCommitments,
     ColumnCommitmentsMismatch, Commitment, DuplicateIdentifiers,
 };
-#[cfg(feature = "arrow")]
-use crate::base::database::{ArrayRefExt, ArrowArrayToColumnConversionError};
 use crate::base::{
-    database::{Column, ColumnField, CommitmentAccessor, OwnedTable, TableRef},
+    database::{ColumnField, CommitmentAccessor, OwnedTable, TableRef},
     scalar::Scalar,
 };
 use alloc::vec::Vec;
-#[cfg(feature = "arrow")]
-use arrow::record_batch::RecordBatch;
-use bumpalo::Bump;
 use core::ops::Range;
-use proof_of_sql_parser::{Identifier, ParseError};
+use proof_of_sql_parser::Identifier;
 use serde::{Deserialize, Serialize};
 use snafu::Snafu;
 
@@ -84,40 +79,40 @@ pub enum TableCommitmentArithmeticError {
 }
 
 /// Errors that can occur when trying to create or extend a [`TableCommitment`] from a record batch.
-#[cfg(feature = "arrow")]
-#[derive(Debug, Snafu)]
-pub enum RecordBatchToColumnsError {
-    /// Error converting from arrow array
-    #[snafu(transparent)]
-    ArrowArrayToColumnConversionError {
-        /// The underlying source error
-        source: ArrowArrayToColumnConversionError,
-    },
-    #[snafu(transparent)]
-    /// This error occurs when convering from a record batch name to an identifier fails. (Which may be impossible.)
-    FieldParseFail {
-        /// The underlying source error
-        source: ParseError,
-    },
-}
+// #[cfg(feature = "arrow")]
+// #[derive(Debug, Snafu)]
+// pub enum RecordBatchToColumnsError {
+//     /// Error converting from arrow array
+//     #[snafu(transparent)]
+//     ArrowArrayToColumnConversionError {
+//         /// The underlying source error
+//         source: ArrowArrayToColumnConversionError,
+//     },
+//     #[snafu(transparent)]
+//     /// This error occurs when convering from a record batch name to an identifier fails. (Which may be impossible.)
+//     FieldParseFail {
+//         /// The underlying source error
+//         source: ParseError,
+//     },
+// }
 
-/// Errors that can occur when attempting to append a record batch to a [`TableCommitment`].
-#[cfg(feature = "arrow")]
-#[derive(Debug, Snafu)]
-pub enum AppendRecordBatchTableCommitmentError {
-    /// During commitment operation, metadata indicates that operand tables cannot be the same.
-    #[snafu(transparent)]
-    ColumnCommitmentsMismatch {
-        /// The underlying source error
-        source: ColumnCommitmentsMismatch,
-    },
-    /// Error converting from arrow array
-    #[snafu(transparent)]
-    ArrowBatchToColumnError {
-        /// The underlying source error
-        source: RecordBatchToColumnsError,
-    },
-}
+// /// Errors that can occur when attempting to append a record batch to a [`TableCommitment`].
+// #[cfg(feature = "arrow")]
+// #[derive(Debug, Snafu)]
+// pub enum AppendRecordBatchTableCommitmentError {
+//     /// During commitment operation, metadata indicates that operand tables cannot be the same.
+//     #[snafu(transparent)]
+//     ColumnCommitmentsMismatch {
+//         /// The underlying source error
+//         source: ColumnCommitmentsMismatch,
+//     },
+//     /// Error converting from arrow array
+//     #[snafu(transparent)]
+//     ArrowBatchToColumnError {
+//         /// The underlying source error
+//         source: RecordBatchToColumnsError,
+//     },
+// }
 
 /// Commitment for an entire table, with column and table metadata.
 ///
@@ -399,89 +394,89 @@ impl<C: Commitment> TableCommitment<C> {
         })
     }
 
-    /// Append an arrow [`RecordBatch`] to the existing [`TableCommitment`].
-    ///
-    /// The row offset is assumed to be the end of the [`TableCommitment`]'s current range.
-    ///
-    /// Will error on a variety of mismatches, or if the provided columns have mixed length.
-    #[cfg(feature = "arrow")]
-    #[allow(clippy::missing_panics_doc)]
-    pub fn try_append_record_batch(
-        &mut self,
-        batch: &RecordBatch,
-        setup: &C::PublicSetup<'_>,
-    ) -> Result<(), AppendRecordBatchTableCommitmentError> {
-        match self.try_append_rows(
-            batch_to_columns::<C::Scalar>(batch, &Bump::new())?
-                .iter()
-                .map(|(a, b)| (a, b)),
-            setup,
-        ) {
-            Ok(()) => Ok(()),
-            Err(AppendTableCommitmentError::MixedLengthColumns { .. }) => {
-                panic!("RecordBatches cannot have columns of mixed length")
-            }
-            Err(AppendTableCommitmentError::AppendColumnCommitments {
-                source: AppendColumnCommitmentsError::DuplicateIdentifiers { .. },
-            }) => {
-                panic!("RecordBatches cannot have duplicate identifiers")
-            }
-            Err(AppendTableCommitmentError::AppendColumnCommitments {
-                source: AppendColumnCommitmentsError::Mismatch { source: e },
-            }) => Err(e)?,
-        }
-    }
-    /// Returns a [`TableCommitment`] to the provided arrow [`RecordBatch`].
-    #[cfg(feature = "arrow")]
-    pub fn try_from_record_batch(
-        batch: &RecordBatch,
-        setup: &C::PublicSetup<'_>,
-    ) -> Result<TableCommitment<C>, RecordBatchToColumnsError> {
-        Self::try_from_record_batch_with_offset(batch, 0, setup)
-    }
+//     /// Append an arrow [`RecordBatch`] to the existing [`TableCommitment`].
+//     ///
+//     /// The row offset is assumed to be the end of the [`TableCommitment`]'s current range.
+//     ///
+//     /// Will error on a variety of mismatches, or if the provided columns have mixed length.
+//     #[cfg(feature = "arrow")]
+//     #[allow(clippy::missing_panics_doc)]
+//     pub fn try_append_record_batch(
+//         &mut self,
+//         batch: &RecordBatch,
+//         setup: &C::PublicSetup<'_>,
+//     ) -> Result<(), AppendRecordBatchTableCommitmentError> {
+//         match self.try_append_rows(
+//             batch_to_columns::<C::Scalar>(batch, &Bump::new())?
+//                 .iter()
+//                 .map(|(a, b)| (a, b)),
+//             setup,
+//         ) {
+//             Ok(()) => Ok(()),
+//             Err(AppendTableCommitmentError::MixedLengthColumns { .. }) => {
+//                 panic!("RecordBatches cannot have columns of mixed length")
+//             }
+//             Err(AppendTableCommitmentError::AppendColumnCommitments {
+//                 source: AppendColumnCommitmentsError::DuplicateIdentifiers { .. },
+//             }) => {
+//                 panic!("RecordBatches cannot have duplicate identifiers")
+//             }
+//             Err(AppendTableCommitmentError::AppendColumnCommitments {
+//                 source: AppendColumnCommitmentsError::Mismatch { source: e },
+//             }) => Err(e)?,
+//         }
+//     }
+//     /// Returns a [`TableCommitment`] to the provided arrow [`RecordBatch`].
+//     #[cfg(feature = "arrow")]
+//     pub fn try_from_record_batch(
+//         batch: &RecordBatch,
+//         setup: &C::PublicSetup<'_>,
+//     ) -> Result<TableCommitment<C>, RecordBatchToColumnsError> {
+//         Self::try_from_record_batch_with_offset(batch, 0, setup)
+//     }
 
-    /// Returns a [`TableCommitment`] to the provided arrow [`RecordBatch`] with the given row offset.
-    #[allow(clippy::missing_panics_doc)]
-    #[cfg(feature = "arrow")]
-    pub fn try_from_record_batch_with_offset(
-        batch: &RecordBatch,
-        offset: usize,
-        setup: &C::PublicSetup<'_>,
-    ) -> Result<TableCommitment<C>, RecordBatchToColumnsError> {
-        match Self::try_from_columns_with_offset(
-            batch_to_columns::<C::Scalar>(batch, &Bump::new())?
-                .iter()
-                .map(|(a, b)| (a, b)),
-            offset,
-            setup,
-        ) {
-            Ok(commitment) => Ok(commitment),
-            Err(TableCommitmentFromColumnsError::MixedLengthColumns { .. }) => {
-                panic!("RecordBatches cannot have columns of mixed length")
-            }
-            Err(TableCommitmentFromColumnsError::DuplicateIdentifiers { .. }) => {
-                panic!("RecordBatches cannot have duplicate identifiers")
-            }
-        }
-    }
-}
+//     /// Returns a [`TableCommitment`] to the provided arrow [`RecordBatch`] with the given row offset.
+//     #[allow(clippy::missing_panics_doc)]
+//     #[cfg(feature = "arrow")]
+//     pub fn try_from_record_batch_with_offset(
+//         batch: &RecordBatch,
+//         offset: usize,
+//         setup: &C::PublicSetup<'_>,
+//     ) -> Result<TableCommitment<C>, RecordBatchToColumnsError> {
+//         match Self::try_from_columns_with_offset(
+//             batch_to_columns::<C::Scalar>(batch, &Bump::new())?
+//                 .iter()
+//                 .map(|(a, b)| (a, b)),
+//             offset,
+//             setup,
+//         ) {
+//             Ok(commitment) => Ok(commitment),
+//             Err(TableCommitmentFromColumnsError::MixedLengthColumns { .. }) => {
+//                 panic!("RecordBatches cannot have columns of mixed length")
+//             }
+//             Err(TableCommitmentFromColumnsError::DuplicateIdentifiers { .. }) => {
+//                 panic!("RecordBatches cannot have duplicate identifiers")
+//             }
+//         }
+//     }
+// }
 
-#[cfg(feature = "arrow")]
-fn batch_to_columns<'a, S: Scalar + 'a>(
-    batch: &'a RecordBatch,
-    alloc: &'a Bump,
-) -> Result<Vec<(Identifier, Column<'a, S>)>, RecordBatchToColumnsError> {
-    batch
-        .schema()
-        .fields()
-        .into_iter()
-        .zip(batch.columns())
-        .map(|(field, array)| {
-            let identifier: Identifier = field.name().parse()?;
-            let column: Column<S> = array.to_column(alloc, &(0..array.len()), None)?;
-            Ok((identifier, column))
-        })
-        .collect()
+// #[cfg(feature = "arrow")]
+// fn batch_to_columns<'a, S: Scalar + 'a>(
+//     batch: &'a RecordBatch,
+//     alloc: &'a Bump,
+// ) -> Result<Vec<(Identifier, Column<'a, S>)>, RecordBatchToColumnsError> {
+//     batch
+//         .schema()
+//         .fields()
+//         .into_iter()
+//         .zip(batch.columns())
+//         .map(|(field, array)| {
+//             let identifier: Identifier = field.name().parse()?;
+//             let column: Column<S> = array.to_column(alloc, &(0..array.len()), None)?;
+//             Ok((identifier, column))
+//         })
+//         .collect()
 }
 
 /// Return the number of rows for the provided columns, erroring if they have mixed length.
@@ -1262,58 +1257,5 @@ mod tests {
             try_negative_low_difference_result,
             Err(TableCommitmentArithmeticError::NegativeRange { .. })
         ));
-    }
-
-    #[test]
-    fn we_can_create_and_append_table_commitments_with_record_batchs() {
-        let batch = record_batch!(
-            "a" => [1i64, 2, 3],
-            "b" => ["1", "2", "3"],
-        );
-
-        let b_scals = ["1".into(), "2".into(), "3".into()];
-
-        let columns = [
-            (
-                &"a".parse().unwrap(),
-                &Column::<Curve25519Scalar>::BigInt(&[1, 2, 3]),
-            ),
-            (
-                &"b".parse().unwrap(),
-                &Column::<Curve25519Scalar>::VarChar((&["1", "2", "3"], &b_scals)),
-            ),
-        ];
-
-        let mut expected_commitment =
-            TableCommitment::<RistrettoPoint>::try_from_columns_with_offset(columns, 0, &())
-                .unwrap();
-
-        let mut commitment =
-            TableCommitment::<RistrettoPoint>::try_from_record_batch(&batch, &()).unwrap();
-
-        assert_eq!(commitment, expected_commitment);
-
-        let batch2 = record_batch!(
-            "a" => [4i64, 5, 6],
-            "b" => ["4", "5", "6"],
-        );
-
-        let b_scals2 = ["4".into(), "5".into(), "6".into()];
-
-        let columns2 = [
-            (
-                &"a".parse().unwrap(),
-                &Column::<Curve25519Scalar>::BigInt(&[4, 5, 6]),
-            ),
-            (
-                &"b".parse().unwrap(),
-                &Column::<Curve25519Scalar>::VarChar((&["4", "5", "6"], &b_scals2)),
-            ),
-        ];
-
-        expected_commitment.try_append_rows(columns2, &()).unwrap();
-        commitment.try_append_record_batch(&batch2, &()).unwrap();
-
-        assert_eq!(commitment, expected_commitment);
     }
 }
