+++ crates/proof-of-sql/src/base/database/mod.rs
@@ -22,14 +22,12 @@ mod table_ref;
 pub use table_ref::TableRef;
 
 #[cfg(feature = "arrow")]
-mod arrow_array_to_column_conversion;
-#[cfg(feature = "arrow")]
-pub use arrow_array_to_column_conversion::{ArrayRefExt, ArrowArrayToColumnConversionError};
-
-#[cfg(feature = "arrow")]
-mod record_batch_utility;
-#[cfg(feature = "arrow")]
-pub use record_batch_utility::ToArrow;
+pub use crate::base::arrow::{
+    arrow_array_to_column_conversion::{ArrayRefExt, ArrowArrayToColumnConversionError},
+    record_batch_utility::ToArrow,
+    owned_and_arrow_conversions::OwnedArrowConversionError,
+    scalar_and_i256_conversions,
+};
 
 #[cfg(all(test, feature = "arrow", feature = "test"))]
 mod test_accessor_utility;
@@ -60,12 +58,6 @@ mod expression_evaluation_error;
 mod expression_evaluation_test;
 pub use expression_evaluation_error::{ExpressionEvaluationError, ExpressionEvaluationResult};
 
-#[cfg(feature = "arrow")]
-mod owned_and_arrow_conversions;
-#[cfg(feature = "arrow")]
-pub use owned_and_arrow_conversions::OwnedArrowConversionError;
-#[cfg(all(test, feature = "arrow"))]
-mod owned_and_arrow_conversions_test;
 
 #[cfg(any(test, feature = "test"))]
 mod test_accessor;
@@ -85,9 +77,6 @@ mod owned_table_test_accessor;
 pub use owned_table_test_accessor::OwnedTableTestAccessor;
 #[cfg(all(test, feature = "blitzar"))]
 mod owned_table_test_accessor_test;
-/// Contains traits for scalar <-> i256 conversions
-#[cfg(feature = "arrow")]
-pub mod scalar_and_i256_conversions;
 
 /// TODO: add docs
 pub(crate) mod filter_util;
