+++ crates/proof-of-sql/src/base/mod.rs
@@ -1,5 +1,6 @@
 //! This module contains basic shared functionalities of the library.
 /// TODO: add docs
+pub mod arrow;
 pub(crate) mod bit;
 pub mod commitment;
 pub mod database;
